requirements.txt
alembic==1.4.3
APScheduler==3.6.3
click==7.1.2
coverage==5.3
dnspython==1.16.0
eventlet==0.29.1
Flask==1.1.2
Flask-APScheduler==1.11.0
Flask-JSGlue==0.3.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==0.4.17
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.4
six==1.15.0
SQLAlchemy==1.3.20
tzlocal==2.1
Werkzeug==1.0.1
WTForms==2.3.3
